Add destroy item
