+++ app/assets/stylesheets/local/gdrive/items.scss
@@ -1,4 +1,4 @@
-.gdrive--settings {
+.gdrive--items {
   ul.permissions {
     margin-bottom: 0;
     padding-inline-start: 18px;
