+++ app/views/gdrive/browse/index.html.erb
@@ -2,7 +2,7 @@
 <% content_for(:action_links) do %>
   <% if @setup_policy.setup? %>
     <% if @config && !@no_credentials && !@authorization_error %>
-      <%= link_to(icon_tag("gear") << " Settings", gdrive_settings_path,
+      <%= link_to(icon_tag("gear") << " Settings", gdrive_items_path,
             class: "btn btn-default") %>
     <% end %>
     <% if FeatureFlag.lookup(:gdrive_migration).on?(current_user) &&
