+++ config/locales/en/en.yml
@@ -96,6 +96,8 @@ en:
       destroy: "Are you sure you want to cancel this event? This action cannot be undone."
     calendars/group:
       destroy: "Are you sure you want to delete this calendar group? This action cannot be undone."
+    gdrive/item:
+      destroy: "Are you sure you want to remove this item? It won't be deleted from Google Drive, but it won't be linked to Gather anymore."
     meals/formula:
       destroy: "Are you sure you want to delete the formula '%{name}'? This action cannot be undone."
       deactivate: "Are you sure you want to deactivate the formula '%{name}'?"
@@ -162,6 +164,9 @@ en:
     billing/template:
       success:
         destroy: "Template deleted successfully."
+    gdrive/item:
+      success:
+        destroy: "Item deleted successfully."
     groups/group:
       success:
         deactivate: "Group deactivated successfully."
